feat:删除部分无效依赖

+++ eros-web/src/main/java/com/github/eros/controller/user/UserInfoHolder.java
@@ -1,7 +1,6 @@
 package com.github.eros.controller.user;
 
-import com.github.eros.entry.UserInfo;
-
+import com.github.eros.domain.user.UserInfo;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
@@ -11,17 +10,17 @@ import java.util.concurrent.ConcurrentHashMap;
  * @date 2021/12/20 02:23
  */
 public class UserInfoHolder {
-    private static final Map<Integer, UserInfo> USER_INFO = new ConcurrentHashMap<>(256);
+    private static final Map<Long, UserInfo> USER_INFO = new ConcurrentHashMap<>(256);
 
     public static void add(UserInfo userInfo) {
         USER_INFO.put(userInfo.getId(), userInfo);
     }
 
-    public static void remove(Integer id) {
+    public static void remove(Long id) {
         USER_INFO.remove(id);
     }
 
-    public static Map<Integer, UserInfo> getAll() {
+    public static Map<Long, UserInfo> getAll() {
         return USER_INFO;
     }
